mv autodiff to interface
