+++ src/Blocks/blockoperations.jl
@@ -99,7 +99,7 @@ expect(op::MatrixBlock, dm::DensityMatrix) = mapslices(x->sum(mat(op).*x)[], dm.
 expect(op::MatrixBlock, dm::DensityMatrix{1}) = sum(mat(op).*dropdims(dm.state, dims=3))
 
 ################### AutoDiff Circuit ###################
-export autodiff, gradient, loss_expect!, loss_Z1!
+export gradient
 """
     gradient(U::AbstractBlock, δ::AbstractRegister)
 
@@ -115,38 +115,3 @@ function gradient(U::AbstractBlock, δ::AbstractRegister)
     grad
 end
 
-"""
-    autodiff(block::AbstractBlock) -> AbstractBlock
-
-automatically mark differentiable items in a block tree as differentiable.
-"""
-function autodiff end
-autodiff(block::Rotor{N}) where N = Diff(block)
-# control, repeat, kron, roller and Diff can not propagate.
-autodiff(block::AbstractBlock) = block
-function autodiff(blk::Union{ChainBlock, Roller, Sequential})
-    chsubblocks(blk, autodiff.(subblocks(blk)))
-end
-
-"""
-    loss_expect(circuit::AbstractBlock, op::AbstractBlock) -> Function
-
-Return function "loss!(ψ, θ) -> Vector"
-"""
-function loss_expect!(circuit::AbstractBlock, op::AbstractBlock)
-    N = nqubits(circuit)
-    function loss!(ψ::AbstractRegister, θ::Vector)
-        params = parameters(circuit)
-        dispatch!(circuit, θ)
-        ψ |> circuit
-        dispatch!!(circuit, params)
-        expect(op, ψ)
-    end
-end
-
-"""
-    loss_Z1!(circuit::AbstractBlock; ibit::Int=1) -> Function
-
-Return the loss function f = <Zi> (means measuring the ibit-th bit in computation basis).
-"""
-loss_Z1!(circuit::AbstractBlock; ibit::Int=1) = loss_expect!(circuit, put(nqubits(circuit), ibit=>Z))
